+++ NEWS.md
@@ -1,3 +1,9 @@
+# bartcs 1.3.0
+
+* Add `synthetic_data()` to create synthetic data for simulation.
+* Add posterior samples of individual potential outcomes.
+* Change `ATE` to `SATE`.
+
 # bartcs 1.2.2
 
 * Remove `vignettes/article.pdf`.